# [Member name lookup] Eliminate non-lazy member loading.