… test/NameLookup/named_lazy_member_loading_objc_category.swift
… test/NameLookup/named_lazy_member_loading_objc_interface.swift
… test/NameLookup/named_lazy_member_loading_objc_protocol.swift
… test/NameLookup/named_lazy_member_loading_protocol_mirroring.swift
… test/NameLookup/named_lazy_member_loading_swift_class.swift
… test/NameLookup/named_lazy_member_loading_swift_class_type.swift
… test/NameLookup/named_lazy_member_loading_swift_derived_class.swift
… test/NameLookup/named_lazy_member_loading_swift_derived_class_type.swift
… test/NameLookup/named_lazy_member_loading_swift_enum.swift
… test/NameLookup/named_lazy_member_loading_swift_proto.swift
… test/NameLookup/named_lazy_member_loading_swift_struct.swift
… test/NameLookup/named_lazy_member_loading_swift_struct_ext.swift
… test/NameLookup/named_lazy_member_loading_swift_struct_ext_mem.swift