@DougGregor 2023-06-01
@swift-ci please test compiler performance