@DougGregor 2023-06-02
@swift-ci please test compiler performance