@DougGregor 2023-06-03
@swift-ci please smoke test compiler performance