@DougGregor 2023-06-05
@swift-ci please smoke test compiler performance