@hamishknight 2023-06-05
Comment from CI run:

# Summary for main full

**Regressions found (see below)**

- [Debug-batch](#debug-batch-3c71233909ecee7f)
    - [debug-batch brief](#debug-batch-brief-3c71233909ecee7f)
    - [debug-batch detailed](#debug-batch-detailed-3c71233909ecee7f)
- [Release](#release-3c71233909ecee7f)
    - [release brief](#release-brief-3c71233909ecee7f)
    - [release detailed](#release-detailed-3c71233909ecee7f)



<a name="debug-batch-3c71233909ecee7f">

# Debug-batch

<a name="debug-batch-brief-3c71233909ecee7f">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 123,168,467,553,071 | 121,896,778,111,005 | -1,271,689,442,065 | -1.03% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
LLVM.NumLLVMBytesOutput | 2,451,467,888 | 2,451,504,720 | 36,832 | 0.0%
</details>

<a name="debug-batch-detailed-3c71233909ecee7f">

## debug-batch detailed
<details>
<summary>Regressed (10)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumSourceLinesPerSecond | 4,804,394 | 4,962,646 | 158,252 | 3.29% :no_entry:
Sema.AccessLevelRequest | 18,991,661 | 19,190,926 | 199,265 | 1.05% :no_entry:
Sema.ClangDirectLookupRequest | 15,213 | 19,430 | 4,217 | 27.72% :no_entry:
Sema.ClangRecordMemberLookup | 15,213 | 19,430 | 4,217 | 27.72% :no_entry:
Sema.CollectOverriddenDeclsRequest | 15,397,304 | 15,575,898 | 178,594 | 1.16% :no_entry:
Sema.LookupAllConformancesInContextRequest | 3,170,796 | 3,204,567 | 33,771 | 1.07% :no_entry:
Sema.NamedLazyMemberLoadSuccessCount | 37,750,592 | 64,542,992 | 26,792,400 | 70.97% :no_entry:
Sema.NumUnloadedLazyIterableDeclContexts | 8,784,224 | 11,187,995 | 2,403,771 | 27.36% :no_entry:
Sema.ProvideDefaultImplForRequest | 15,397,304 | 15,575,898 | 178,594 | 1.16% :no_entry:
Sema.USRGenerationRequest | 17,144,407 | 17,329,400 | 184,993 | 1.08% :no_entry:
</details>
<details>
<summary>Improved (15)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumASTBytesAllocated | 66,278,458,904 | 60,393,996,855 | -5,884,462,049 | -8.88% :white_check_mark:
AST.NumTotalClangImportedEntities | 1,961,936 | 1,812,391 | -149,545 | -7.62% :white_check_mark:
Frontend.NumInstructionsExecuted | 123,168,467,553,071 | 121,896,778,111,005 | -1,271,689,442,065 | -1.03% :white_check_mark:
Sema.CxxRecordSemantics | 85,742 | 71,810 | -13,931 | -16.25% :white_check_mark:
Sema.ExpandMemberAttributeMacros | 72,347,779 | 65,464,830 | -6,882,948 | -9.51% :white_check_mark:
Sema.GenericSignatureRequest | 3,639,277 | 3,599,834 | -39,443 | -1.08% :white_check_mark:
Sema.InheritedProtocolsRequest | 866,596 | 854,151 | -12,444 | -1.44% :white_check_mark:
Sema.LookupInModuleRequest | 10,043,683 | 9,697,853 | -345,830 | -3.44% :white_check_mark:
Sema.ModuleQualifiedLookupRequest | 4,065,177 | 3,718,818 | -346,359 | -8.52% :white_check_mark:
Sema.NumDeclsDeserialized | 120,070,444 | 97,747,957 | -22,322,487 | -18.59% :white_check_mark:
Sema.NumLazyIterableDeclContexts | 15,574,409 | 14,701,592 | -872,817 | -5.6% :white_check_mark:
Sema.NumLazyRequirementSignatures | 1,191,761 | 1,126,780 | -64,980 | -5.45% :white_check_mark:
Sema.NumRequirementMachineCompletionSteps | 29,942,759 | 27,562,046 | -2,380,712 | -7.95% :white_check_mark:
Sema.NumTypesDeserialized | 32,893,865 | 27,190,177 | -5,703,688 | -17.34% :white_check_mark:
Sema.SPIGroupsRequest | 8,823,278 | 8,701,143 | -122,135 | -1.38% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (298)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 3,853,839 | 3,845,099 | -8,740 | -0.23%
AST.ImportSetCacheMiss | 794,931 | 793,334 | -1,596 | -0.2%
AST.ImportSetFoldHit | 403,232 | 402,697 | -534 | -0.13%
AST.ImportSetFoldMiss | 391,699 | 390,637 | -1,062 | -0.27%
AST.ModuleShadowCacheHit | 54,516 | 54,486 | -30 | -0.06%
AST.ModuleShadowCacheMiss | 4,395 | 4,402 | 7 | 0.16%
AST.ModuleVisibilityCacheHit | 27,745,092 | 27,745,093 | 0 | 0.0%
AST.ModuleVisibilityCacheMiss | 209,911 | 209,912 | 0 | 0.0%
AST.NumASTScopeExpansions | 8,360,754 | 8,361,002 | 248 | 0.0%
AST.NumASTScopeLookups | 9,575,859 | 9,576,558 | 699 | 0.01%
AST.NumDecls | 342,185 | 342,185 | 0 | 0.0%
AST.NumDependencies | 431,187 | 431,187 | 0 | 0.0%
AST.NumIncrementalDependencies | 25,214 | 25,214 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 323,059 | 323,059 | 0 | 0.0%
AST.NumLocalTypeDecls | 6,522 | 6,522 | 0 | 0.0%
AST.NumModuleLookupClassMember | 6,798 | 6,798 | 0 | 0.0%
AST.NumModuleLookupValue | 114,237,641 | 113,799,968 | -437,672 | -0.38%
AST.NumObjCMethods | 93,976 | 93,976 | 0 | 0.0%
AST.NumOperators | 1,028 | 1,028 | 0 | 0.0%
AST.NumPrecedenceGroups | 176 | 176 | 0 | 0.0%
AST.NumReferencedDynamicNames | 159 | 159 | 0 | 0.0%
AST.NumReferencedMemberNames | 15,011,372 | 15,006,462 | -4,909 | -0.03%
AST.NumReferencedTopLevelNames | 1,349,062 | 1,341,662 | -7,400 | -0.55%
AST.NumSourceBuffers | 396,328 | 395,235 | -1,093 | -0.28%
AST.NumSourceLines | 12,407,592 | 12,407,592 | 0 | 0.0%
Frontend.MaxMallocUsage | 0 | 0 | 0 | 0.0%
Frontend.NumProcessFailures | 31 | 31 | 0 | 0.0%
IRGen.IRGenRequest | 29,562 | 29,562 | 0 | 0.0%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 298,484 | 298,484 | 0 | 0.0%
IRModule.NumIRAliases | 240,381 | 240,381 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 8,282,853 | 8,282,853 | 0 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 4,510,403 | 4,510,403 | 0 | 0.0%
IRModule.NumIRGlobals | 4,284,206 | 4,284,206 | 0 | 0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 105,840,628 | 105,840,628 | 0 | 0.0%
IRModule.NumIRNamedMetaData | 119,768 | 119,768 | 0 | 0.0%
IRModule.NumIRValueSymbols | 8,866,897 | 8,866,897 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 2,451,467,888 | 2,451,504,720 | 36,832 | 0.0%
Parse.ExportedSourceFileRequest | 0 | 0 | 0 | 0.0%
Parse.IDEInspectionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 398,787 | 398,787 | 0 | 0.0%
Parse.NumIterableDeclContextParsed | 808,274 | 808,294 | 20 | 0.0%
Parse.ParseAbstractFunctionBodyRequest | 687,515 | 687,515 | 0 | 0.0%
Parse.ParseMembersRequest | 619,453 | 619,473 | 20 | 0.0%
Parse.ParseSourceFileRequest | 370,601 | 370,601 | 0 | 0.0%
Parse.ParseTopLevelDeclsRequest | 370,601 | 370,601 | 0 | 0.0%
SILGen.ASTLoweringRequest | 30,898 | 30,898 | 0 | 0.0%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 3,527,077 | 3,527,077 | 0 | 0.0%
SILModule.NumSILGenGlobalVariables | 141,265 | 141,265 | 0 | 0.0%
SILModule.NumSILGenVtables | 40,409 | 40,409 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 198,405 | 198,405 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 4,030,010 | 4,030,010 | 0 | 0.0%
SILModule.NumSILOptGlobalVariables | 143,293 | 143,293 | 0 | 0.0%
SILModule.NumSILOptVtables | 40,772 | 40,772 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 242,138 | 242,138 | 0 | 0.0%
SILOptimizer.ExecuteSILPipelineRequest | 151,846 | 151,846 | 0 | 0.0%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 310,173 | 310,175 | 1 | 0.0%
Sema.AbstractGenericSignatureRequest | 146,289 | 146,652 | 363 | 0.25%
Sema.ActorIsolationRequest | 7,141,949 | 7,143,454 | 1,505 | 0.02%
Sema.AllMembersRequest | 118,246 | 118,246 | 0 | 0.0%
Sema.AnyObjectLookupRequest | 234 | 234 | 0 | 0.0%
Sema.ApplyAccessNoteRequest | 4,386,517 | 4,387,626 | 1,109 | 0.03%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 49,679 | 49,679 | 0 | 0.0%
Sema.AttachedPropertyWrapperTypeRequest | 30,650 | 30,650 | 0 | 0.0%
Sema.AttachedPropertyWrappersRequest | 23,120,012 | 23,265,465 | 145,453 | 0.63%
Sema.AttachedResultBuilderRequest | 1,890,539 | 1,890,539 | 0 | 0.0%
Sema.BodyInitKindRequest | 94,561 | 94,561 | 0 | 0.0%
Sema.BreakTargetRequest | 6,644 | 6,644 | 0 | 0.0%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 480,101 | 480,101 | 0 | 0.0%
Sema.CheckDistributedFunctionRequest | 80 | 80 | 0 | 0.0%
Sema.CheckInconsistentAccessLevelOnImport | 59,646 | 59,646 | 0 | 0.0%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 10,772 | 10,772 | 0 | 0.0%
Sema.CheckInconsistentSPIOnlyImportsRequest | 59,646 | 59,646 | 0 | 0.0%
Sema.CheckInconsistentWeakLinkedImportsRequest | 10,772 | 10,772 | 0 | 0.0%
Sema.CheckRedeclarationRequest | 2,343,215 | 2,343,215 | 0 | 0.0%
Sema.ClangCategoryLookupRequest | 0 | 0 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 153,270 | 153,381 | 111 | 0.07%
Sema.ClosureEffectsRequest | 153,396 | 153,396 | 0 | 0.0%
Sema.ClosureHasExplicitResultRequest | 58,921 | 58,921 | 0 | 0.0%
Sema.CompareDeclSpecializationRequest | 875,481 | 876,754 | 1,273 | 0.15%
Sema.CompilerPluginLoadRequest | 0 | 0 | 0 | 0.0%
Sema.ConditionalRequirementsRequest | 880,162 | 881,868 | 1,705 | 0.19%
Sema.ConformanceHasEffectRequest | 63 | 63 | 0 | 0.0%
Sema.ConstantValueInfoRequest | 0 | 0 | 0 | 0.0%
Sema.ContinueTargetRequest | 2,182 | 2,182 | 0 | 0.0%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 22,083 | 22,083 | 0 | 0.0%
Sema.CustomAttrTypeRequest | 3,280 | 3,280 | 0 | 0.0%
Sema.CustomRefCountingOperation | 0 | 0 | 0 | 0.0%
Sema.CxxRecordAsSwiftType | 0 | 0 | 0 | 0.0%
Sema.DefaultAndMaxAccessLevelRequest | 121,948 | 121,948 | 0 | 0.0%
Sema.DefaultArgumentExprRequest | 103,166 | 103,166 | 0 | 0.0%
Sema.DefaultArgumentInitContextRequest | 841 | 841 | 0 | 0.0%
Sema.DefaultArgumentTypeRequest | 21,729 | 21,729 | 0 | 0.0%
Sema.DefaultDefinitionTypeRequest | 8,514 | 8,514 | 0 | 0.0%
Sema.DefaultTypeRequest | 656,698 | 656,698 | 0 | 0.0%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 81,289,643 | 81,348,950 | 59,307 | 0.07%
Sema.DirectOperatorLookupRequest | 2,586,065 | 2,580,041 | -6,023 | -0.23%
Sema.DirectPrecedenceGroupLookupRequest | 953,005 | 952,854 | -151 | -0.02%
Sema.DistributedModuleIsAvailableRequest | 67 | 67 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 7,591,132 | 7,592,637 | 1,505 | 0.02%
Sema.EnumElementExprPatternRequest | 17 | 17 | 0 | 0.0%
Sema.EnumRawTypeRequest | 41,463 | 41,463 | 0 | 0.0%
Sema.EnumRawValuesRequest | 24,380 | 24,380 | 0 | 0.0%
Sema.ExistentialConformsToSelfRequest | 33,932 | 33,964 | 32 | 0.1%
Sema.ExpandAccessorMacros | 1,308,294 | 1,308,294 | 0 | 0.0%
Sema.ExpandConformanceMacros | 1,297,265 | 1,303,330 | 6,065 | 0.47%
Sema.ExpandMacroExpansionDeclRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMacroExpansionExprRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandPeerMacroRequest | 10,293,121 | 10,293,123 | 2 | 0.0%
Sema.ExpandSynthesizedMemberMacroRequest | 304,165 | 304,167 | 1 | 0.0%
Sema.ExprPatternMatchRequest | 93,117 | 93,117 | 0 | 0.0%
Sema.ExtendedNominalRequest | 708,143 | 708,143 | 0 | 0.0%
Sema.ExtendedTypeRequest | 217,776 | 217,776 | 0 | 0.0%
Sema.ExternalMacroDefinitionRequest | 0 | 0 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 2,802,822 | 2,786,900 | -15,922 | -0.57%
Sema.FunctionOperatorRequest | 20,484 | 20,484 | 0 | 0.0%
Sema.GenericParamListRequest | 7,413,957 | 7,352,322 | -61,634 | -0.83%
Sema.GetDestructorRequest | 47,217 | 47,217 | 0 | 0.0%
Sema.GetDistributedActorArgumentDecodingMethodRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 72 | 72 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 117 | 117 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 63 | 63 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 46 | 46 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 22 | 22 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 20 | 20 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 11 | 11 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 16 | 16 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 12 | 12 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 80 | 80 | 0 | 0.0%
Sema.GetImplicitSendableRequest | 264,389 | 264,997 | 607 | 0.23%
Sema.GetRuntimeDiscoverableAttributes | 2,464,712 | 2,464,712 | 0 | 0.0%
Sema.GetSourceFileAsyncNode | 138 | 138 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 9,071,054 | 9,073,341 | 2,286 | 0.03%
Sema.GlobalActorInstanceRequest | 1,063 | 1,063 | 0 | 0.0%
Sema.HasCircularInheritedProtocolsRequest | 18,567 | 18,567 | 0 | 0.0%
Sema.HasCircularRawValueRequest | 25,158 | 25,158 | 0 | 0.0%
Sema.HasDefaultInitRequest | 107,257 | 107,257 | 0 | 0.0%
Sema.HasDynamicCallableAttributeRequest | 31,859 | 31,859 | 0 | 0.0%
Sema.HasDynamicMemberLookupAttributeRequest | 582,418 | 582,966 | 548 | 0.09%
Sema.HasImportsMatchingFlagRequest | 735,192 | 735,192 | 0 | 0.0%
Sema.HasIsolatedSelfRequest | 16,043,745 | 16,109,072 | 65,327 | 0.41%
Sema.HasMemberwiseInitRequest | 52,552 | 52,552 | 0 | 0.0%
Sema.HasMissingDesignatedInitializersRequest | 33,351 | 33,351 | 0 | 0.0%
Sema.HasSelfOrAssociatedTypeRequirementsRequest | 16,786 | 16,786 | 0 | 0.0%
Sema.HasUserDefinedDesignatedInitRequest | 107,267 | 107,267 | 0 | 0.0%
Sema.IDEInspectionFileRequest | 0 | 0 | 0 | 0.0%
Sema.ImplementsAttrProtocolRequest | 0 | 0 | 0 | 0.0%
Sema.InferredGenericSignatureRequest | 243,199 | 243,200 | 1 | 0.0%
Sema.InheritedDeclsReferencedRequest | 7,885,259 | 7,907,430 | 22,171 | 0.28%
Sema.InheritedTypeRequest | 476,333 | 476,242 | -91 | -0.02%
Sema.InheritsSuperclassInitializersRequest | 42,298 | 42,298 | 0 | 0.0%
Sema.InitKindRequest | 197,905 | 197,905 | 0 | 0.0%
Sema.InterfaceTypeRequest | 24,654,359 | 24,631,707 | -22,652 | -0.09%
Sema.IsABICompatibleOverrideRequest | 323,615 | 323,615 | 0 | 0.0%
Sema.IsAccessorTransparentRequest | 579,407 | 579,407 | 0 | 0.0%
Sema.IsActorRequest | 2,525,587 | 2,512,594 | -12,992 | -0.51%
Sema.IsCallAsFunctionNominalRequest | 6,761 | 6,761 | 0 | 0.0%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 72,243 | 72,243 | 0 | 0.0%
Sema.IsDefaultActorRequest | 63,999 | 63,999 | 0 | 0.0%
Sema.IsDistributedActorRequest | 2,811,669 | 2,799,664 | -12,005 | -0.43%
Sema.IsDynamicRequest | 3,301,947 | 3,301,947 | 0 | 0.0%
Sema.IsFinalRequest | 5,037,107 | 5,059,271 | 22,164 | 0.44%
Sema.IsGetterMutatingRequest | 831,312 | 831,312 | 0 | 0.0%
Sema.IsImplicitlyUnwrappedOptionalRequest | 4,096,708 | 4,079,616 | -17,092 | -0.42%
Sema.IsMoveOnlyRequest | 784,993 | 780,737 | -4,256 | -0.54%
Sema.IsNonUserModuleRequest | 104,567 | 104,578 | 11 | 0.01%
Sema.IsObjCRequest | 3,074,269 | 3,075,378 | 1,109 | 0.04%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 745,767 | 745,767 | 0 | 0.0%
Sema.IsSingleValueStmtRequest | 375,400 | 375,400 | 0 | 0.0%
Sema.IsStaticRequest | 1,670,463 | 1,670,423 | -39 | -0.0%
Sema.LazyStoragePropertyRequest | 2,573 | 2,573 | 0 | 0.0%
Sema.LocalDiscriminatorsRequest | 555,030 | 555,030 | 0 | 0.0%
Sema.LookupConformanceInModuleRequest | 128,999,956 | 129,130,781 | 130,824 | 0.1%
Sema.LookupInfixOperatorRequest | 82,647 | 82,647 | 0 | 0.0%
Sema.LookupPostfixOperatorRequest | 74 | 74 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 33,630 | 33,630 | 0 | 0.0%
Sema.LookupPrefixOperatorRequest | 454 | 454 | 0 | 0.0%
Sema.MacroDefinitionRequest | 0 | 0 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 3,149 | 3,149 | 0 | 0.0%
Sema.ModuleImplicitImportsRequest | 10,772 | 10,772 | 0 | 0.0%
Sema.ModuleLibraryLevelRequest | 22 | 22 | 0 | 0.0%
Sema.NamingPatternRequest | 242,940 | 242,941 | 1 | 0.0%
Sema.NeedsNewVTableEntryRequest | 1,076,375 | 1,076,375 | 0 | 0.0%
Sema.NumAccessorBodiesSynthesized | 482,841 | 482,841 | 0 | 0.0%
Sema.NumAccessorsSynthesized | 582,107 | 582,107 | 0 | 0.0%
Sema.NumConformancePathsRecorded | 1,000,130 | 1,000,652 | 522 | 0.05%
Sema.NumConformancesDeserialized | 6,280,432 | 6,286,669 | 6,237 | 0.1%
Sema.NumConstraintScopes | 80,484,273 | 80,492,662 | 8,388 | 0.01%
Sema.NumConstraintsConsideredForEdgeContraction | 2,726,882 | 2,726,882 | 0 | 0.0%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsTypechecked | 3,329,446 | 3,329,446 | 0 | 0.0%
Sema.NumLazyRequirementSignaturesLoaded | 792,250 | 791,075 | -1,174 | -0.15%
Sema.NumLeafScopes | 61,952,401 | 61,959,416 | 7,015 | 0.01%
Sema.NumRequirementMachineUnifiedConcreteTerms | 1,362,459 | 1,367,580 | 5,121 | 0.38%
Sema.NumRequirementMachines | 2,345,579 | 2,334,978 | -10,601 | -0.45%
Sema.ObjCInterfaceAndImplementationRequest | 950,925 | 950,925 | 0 | -0.0%
Sema.OpaqueReadOwnershipRequest | 505,180 | 505,180 | 0 | 0.0%
Sema.OpaqueResultTypeRequest | 513 | 513 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 1,027 | 1,027 | 0 | 0.0%
Sema.OverriddenDeclsRequest | 6,690,281 | 6,709,271 | 18,989 | 0.28%
Sema.ParamSpecifierRequest | 2,257,252 | 2,257,252 | 0 | 0.0%
Sema.PatternBindingEntryRequest | 821,524 | 821,525 | 1 | 0.0%
Sema.PatternTypeRequest | 974,136 | 974,137 | 1 | 0.0%
Sema.PolymorphicEffectKindRequest | 1,125,091 | 1,125,091 | 0 | 0.0%
Sema.PolymorphicEffectRequirementsRequest | 75 | 75 | 0 | 0.0%
Sema.PotentialMacroExpansionsInContextRequest | 7,363,465 | 7,369,615 | 6,150 | 0.08%
Sema.PreCheckResultBuilderRequest | 951 | 951 | 0 | 0.0%
Sema.PreCheckReturnStmtRequest | 597,491 | 597,491 | 0 | 0.0%
Sema.PrimaryAssociatedTypesRequest | 19,528 | 19,691 | 163 | 0.84%
Sema.PrimarySourceFilesRequest | 10,772 | 10,772 | 0 | 0.0%
Sema.PropertyWrapperAuxiliaryVariablesRequest | 2,464,995 | 2,464,995 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyTypeRequest | 30,650 | 30,650 | 0 | 0.0%
Sema.PropertyWrapperInitializerInfoRequest | 719,150 | 719,150 | 0 | 0.0%
Sema.PropertyWrapperLValuenessRequest | 2,636 | 2,636 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 825,157 | 825,157 | 0 | 0.0%
Sema.PropertyWrapperTypeInfoRequest | 599 | 599 | 0 | 0.0%
Sema.ProtocolDependenciesRequest | 678,927 | 676,792 | -2,135 | -0.31%
Sema.ProtocolRequirementsRequest | 81,572 | 81,574 | 1 | 0.0%
Sema.ProtocolRequiresClassRequest | 48,247 | 48,135 | -111 | -0.23%
Sema.QualifiedLookupRequest | 10,086,825 | 10,098,514 | 11,689 | 0.12%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RawCommentRequest | 1,193,191 | 1,193,191 | 0 | 0.0%
Sema.RenamedDeclRequest | 5,222 | 5,222 | 0 | 0.0%
Sema.RequirementRequest | 194,305 | 194,305 | 0 | 0.0%
Sema.RequirementSignatureRequest | 852,071 | 850,821 | -1,250 | -0.15%
Sema.RequiresOpaqueAccessorsRequest | 2,109,314 | 2,109,314 | 0 | 0.0%
Sema.RequiresOpaqueModifyCoroutineRequest | 510,431 | 510,431 | 0 | 0.0%
Sema.ResolveEffectiveMemberwiseInitRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveImplicitMemberRequest | 952,509 | 953,738 | 1,229 | 0.13%
Sema.ResolveMacroRequest | 24,225 | 24,225 | 0 | 0.0%
Sema.ResolveProtocolNameRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeEraserTypeRequest | 0 | 0 | 0 | 0.0%
…